requirements.txt
# Core dependencies
requests>=2.28.0

# Document processing
pytesseract>=0.3.10
pdf2image>=1.16.0
pdfplumber>=0.9.0
Pillow>=9.0.0
python-magic>=0.4.27

# NLP and ML
spacy>=3.5.0
scikit-learn>=1.2.0

# AI Detection - Local Models (Required for local detection)
transformers>=4.30.0
torch>=2.0.0

# ROUGE metrics (for pyrouge/similarity detection)
rouge-score>=0.1.2

# For Binoculars detector (requires ~28GB RAM for Falcon-7B models)
accelerate>=0.20.0
scipy

# Optional: For 8-bit quantization (reduces Binoculars memory to ~14GB)
# bitsandbytes>=0.39.0

# Note: API-based detectors only require 'requests' which is included above.
# Set environment variables for API keys:
#   GPTZERO_API_KEY      - https://gptzero.me
#   ORIGINALITY_API_KEY  - https://originality.ai
#   ZEROGPT_API_KEY      - https://zerogpt.com
#   COPYLEAKS_EMAIL + COPYLEAKS_API_KEY - https://copyleaks.com
